requirements.txt
arabic-reshaper
asgiref
asn1crypto
asttokens
backcall
Brotli
certifi
cffi
chardet
charset-normalizer
click
colorama
comm
crispy-bootstrap5
crispy-tailwind
cryptography
cssselect2
debugpy
decorator
distlib
dj-database-url
Django
django-allauth
django-crispy-forms
django-email
django-environ
django-filter
django-rest-auth
djangorestframework
djangorestframework-simplejwt
EasyProcess
entrypoint2
executing
filelock
fonttools
html5lib
idna
ipykernel
ipython
jedi
jupyter_client
jupyter_core
lxml
Markdown
matplotlib-inline
MouseInfo
mss
mysql-connector-python
nest-asyncio
numpy
opencv-python
oscrypto
packaging
parso
pickleshare
Pillow
pipenv
platformdirs
prompt-toolkit
psutil
psycopg2
psycopg2-binary
pure-eval
PyAutoGUI
pycparser
pydyf
PyGetWindow
Pygments
pyHanko
pyhanko-certvalidator
PyJWT
PyMsgBox
PyMySQL
pypdf
pyperclip
pyphen
pypng
PyRect
pyscreenshot
PyScreeze
python-bidi
python-dateutil
pytweening
PyYAML
pyzmq
qrcode
reportlab
requests
six
sqlparse
stack-data
svglib
tinycss2
tornado
traitlets
typing_extensions
tzdata
tzlocal
uritools
urllib3
virtualenv
wcwidth
webencodings
xhtml2pdf
zopfli